Basic Enemy Finite State Machine mock-up completed
Basically, I made a super basic AI system. Right now, the enemy enters
an "idle" state when it's created, moves immediately to an "attack"
state, with an attack that lasts 5 seconds, then goes back to "idle",
which loops. This is easily expandable to more states with behaviors
linked to these states.